requirements.txt
django==4.2.3
djangorestframework==3.14.0
psycopg2==2.8.6
drf-spectacular==0.26.3